+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Added getInlineViewAssets() to Verifier
+* Improved inline rendering
+* Added getName() to Verifier
+
 ## v0.1.0 (2024-06-19)
 * Built initial Verifier implementation
 * Built basic Renderer system